backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
